+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ## v0.1.0 (18/10/2024)
 
-- First release of `pycounts_labdmitriy`!
+- First release of `pycounts-labdmitriy`!